# internal/domains/user/dto/request.go
package dto

import _ "github.com/go-playground/validator/v10"

type UpsertUserRequest struct {
	Name     string `json:"name" validate:"required"`
	Username string `json:"username" validate:"required"`
	Email    string `json:"email" validate:"required,email"`
	Password string `json:"password" validate:"required"`
}